Rename GOVUK::Error to GovukError
While GOVUK::Error is a doper namespace, and conforms more to the
GOV.UK style guide, it's not consistent with other GOV.UK gems like
GovukSidekiq.
