+++ lib/govuk_app_config.rb
@@ -1,4 +1,4 @@
 require "govuk_app_config/version"
-require "govuk_app_config/statsd"
-require "govuk_app_config/error"
+require "govuk_app_config/govuk_statsd"
+require "govuk_app_config/govuk_error"
 require "govuk_app_config/auto_configure"
