+++ lib/govuk_app_config/auto_configure.rb
@@ -9,16 +9,16 @@ Raven.configure do |config|
   # We're misusing the `should_capture` block here to hook into raven until
   # there's a better way: https://github.com/getsentry/raven-ruby/pull/750
   config.should_capture = Proc.new {
-    GOVUK::Statsd.increment("errors_occurred")
+    GovukStatsd.increment("errors_occurred")
 
     # For backwards compatibility
-    GOVUK::Statsd.increment("errbit.errors_occurred")
+    GovukStatsd.increment("errbit.errors_occurred")
 
     # Return true so that we don't accidentally skip the error
     true
   }
 
   config.transport_failure_callback = Proc.new {
-    GOVUK::Statsd.increment("error_reports_failed")
+    GovukStatsd.increment("error_reports_failed")
   }
 end
