+++ spec/lib/govuk_statsd_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
-require 'govuk_app_config/statsd'
+require 'govuk_app_config/govuk_statsd'
 
-RSpec.describe GOVUK::Statsd do
+RSpec.describe GovukStatsd do
   describe '#increment' do
     it 'increments the counter' do
       expect {
-        GOVUK::Statsd.increment("some.key")
+        GovukStatsd.increment("some.key")
       }.not_to raise_error
     end
   end
